requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
pandas==2.3.2
openai
python-socketio==5.9.0
eventlet==0.33.3
xlsxwriter==3.1.6
requests==2.31.0